@@ -13,14 +13,14 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>8</source>
-                    <target>8</target>
+                    <source>21</source>
+                    <target>21</target>
                 </configuration>
             </plugin>
         </plugins>
     </build>
     <properties>
-        <java-version>1.8</java-version>
+        <java-version>21</java-version>
     </properties>
     <dependencies>
         <dependency>
