@@ -1,5 +1,6 @@
 package com;
 
+import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -8,20 +9,23 @@ public class ReverseString {
 
     public static void main(String[] args) {
         String s = String.valueOf(a);
-        s.toLowerCase();
         String regex = "(.)*(\\\\d)(.)*";
         Pattern pattern = Pattern.compile(regex);
         Matcher matcher = pattern.matcher(s);
 
-        String reverse = "";
+        StringBuilder reverse = new StringBuilder();
         for (int i = s.length() - 1; i >= 0; i--) {
-            reverse += s.charAt(i);
+            reverse.append(s.charAt(i));
         }
 
         if (matcher.matches()) {
             System.out.println("matched");
         }
+
+        Optional<String> res = Optional.of(null);
+
+        System.out.println(res.orElse(null));
         System.out.println(reverse);
-        System.out.println(s.equalsIgnoreCase(reverse));
+        System.out.println(s.equalsIgnoreCase(reverse.toString()));
     }
 }
