@@ -8,13 +8,17 @@ import java.util.concurrent.Semaphore;
 
 public class Sema {
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws InterruptedException {
 
         ExecutorService service = Executors.newFixedThreadPool(4);
         Semaphore semaphore = new Semaphore(1);
         for (int i = 0; i < 10; i++) {
-            service.execute(new SemaRun(semaphore, i));
+//            service.execute(new SemaRun(semaphore, i));
+            Thread.startVirtualThread(new SemaRun(semaphore, i));
         }
+
+        Thread.startVirtualThread(() -> System.out.println("I am Virtual Thread"));
+        Thread.sleep(1000);
         service.shutdown();
     }
 }
@@ -34,8 +38,8 @@ class SemaRun implements Runnable {
     public void run() {
         System.out.println("Thread " + atomicInteger);
         sem.acquireUninterruptibly();
-        System.out.println("tad".equals(null));
-        System.out.println("Thread Name " + Thread.currentThread().getId());
+        System.out.println("tad" == null);
+        System.out.println("Thread Name " + Thread.currentThread() + " " + atomicInteger);
         sem.release();
 
     }
