@@ -4,38 +4,59 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Subsequence {
-    static List<String> list = new ArrayList<>();
+    static List<Integer> list = new ArrayList<>();
+    static int count = 0;
 
     public static void main(String[] args) {
-        subSequence("abc", "");
-//        permutations("abc", "");
-        System.out.println(list);
+
+        int[] arr = {3, 2, -1};
+
+//        array(0, arr, arr.length);
+        array(0, arr, arr.length);
+        List<Integer> list1 = new ArrayList<>();
+        int sum = 0, k = 5;
+        sum(arr, sum, arr.length - 1, list1, k);
+        System.out.println(sum(arr, sum, arr.length - 1, k));
+        System.out.println(count);
 
     }
 
-    public static void subSequence(String s, String ans) {
-        if (s.isEmpty()) {
-            list.add(ans);
+
+    /*
+     print all subsequences
+     */
+    private static void array(int i, int[] arr, int n) {
+        if (i >= n) {
+            System.out.println(list);
             return;
         }
-        subSequence(s.substring(1), ans);
-        subSequence(s.substring(1), ans + s.charAt(0));
+        list.add(arr[i]);
+        array(i + 1, arr, n);
+        list.remove(list.size() - 1);
+        array(i + 1, arr, n);
 
-//        System.out.println(ans);
     }
 
-    private static void permutations(String s, String ans) {
-
-        if (s.isEmpty()) list.add(ans);
-
-        for (int i = 0; i < s.length(); i++) {
-            char ch = s.charAt(i);
-
-            // Rest of the string after excluding
-            // the ith character
-            String ros = s.substring(0, i) + s.substring(i + 1);
-
-            permutations(ros, ans + ch);
+    //print all the numbers of sum k
+    private static void sum(int[] arr, int sum, int n, List<Integer> list, int k) {
+        if (sum == k) {
+            System.out.println(list);
+            return;
         }
+        if (n < 0) return;
+        list.add(arr[n]);
+        sum(arr, sum + arr[n], n - 1, list, k);
+        list.remove(list.size() - 1);
+        sum(arr, sum, n - 1, list, k);
+    }
+
+    //count of all sum k in an array
+    private static int sum(int[] arr, int sum, int n, int k) {
+        if (sum == k) {
+            count++;
+            return count;
+        }
+        if (n < 0 || sum > k) return 0;
+        return Math.max(sum(arr, sum + arr[n], n - 1, k), sum(arr, sum, n - 1, k));
     }
 }
